+++ web/src/components/recipe_selection.rs
@@ -68,7 +68,7 @@ pub fn RecipeSelection<'ctx, G: Html>(
         label(for=for_id, class="flex-item-grow") { a(href=href) { (*title) } }
         NumberField(name=name, class="flex-item-shrink".to_string(), counter=count, min=0.0, on_change=Some(move |_| {
             debug!(idx=%id, count=%(*count.get_untracked()), "setting recipe count");
-            sh.dispatch(cx, Message::UpdateRecipeCount(id.as_ref().clone(), *count.get_untracked() as usize));
+            sh.dispatch(cx, Message::UpdateRecipeCount(id.as_ref().clone(), *count.get_untracked() as u32));
         }))
     }
 }
